requirements.txt
anthropic>=0.25.0
chromadb>=0.5.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
